Use proper appdata paths on Windows/macOS
