+++ crates/jackdaw_env/Cargo.toml
@@ -7,6 +7,7 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/jbuehler23/jackdaw"
 
 [dependencies]
+dirs.workspace = true
 
 [lints]
 workspace = true
